examples/requirements.txt
Django>=1.7
django-filer
django-mptt
easy_thumbnails
django-polymorphic
